% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iplotMap.R
\name{iplotMap}
\alias{iplotMap}
\title{Interactive genetic map plot}
\usage{
iplotMap(map, shift = FALSE, chartOpts = NULL)
}
\arguments{
\item{map}{Object of class \code{"map"}, a list with each component
being a vector of marker positions.}

\item{shift}{If TRUE, shift each chromsome so that the initial marker
is at position 0.}

\item{chartOpts}{A list of options for configuring the chart.  Each
  element must be named using the corresponding option.}
}
\value{
None.
}
\description{
Creates an interactive graph of a genetic marker map.
}
\examples{
library(qtl)
data(hyper)
map <- pull.map(hyper)
\donttest{
iplotMap(map, shift=TRUE)}
}
\seealso{
\code{\link{iplotScanone}}, \code{\link{iplotPXG}}
}
\keyword{hplot}

