% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itimeplot.R
\name{itimeplot}
\alias{itimeplot}
\title{Interactive scatterplot with date/time on x-axis}
\usage{
itimeplot(
  datetime,
  y,
  group = NULL,
  indID = NULL,
  chartOpts = NULL,
  digits = 5
)
}
\arguments{
\item{datetime}{Vector of date/time values}

\item{y}{Numeric vector of y values}

\item{group}{Optional vector of categories for coloring the points}

\item{indID}{Optional vector of character strings, shown with tool tips}

\item{chartOpts}{A list of options for configuring the chart.  Each
element must be named using the corresponding option.}

\item{digits}{Round data to this number of significant digits
before passing to the chart function. (Use NULL to not round.)}
}
\value{
An object of class \code{htmlwidget} that will
intelligently print itself into HTML in a variety of contexts
including the R console, within R Markdown documents, and within
Shiny output bindings.
}
\description{
Creates an interactive scatterplot with date/time on the x-axis.
}
\examples{
n_pts <- 100
x <-  seq(as.POSIXct("1969-05-01"), as.POSIXct("1969-05-04"), length=n_pts)
grp <- sample(1:3, n_pts, replace=TRUE)
y <- rnorm(n_pts, grp) + rnorm(n_pts)
\donttest{
itimeplot(x, y, grp)
}

# with shorter time span, times are shown instead of dates
x <-  seq(as.POSIXct("1969-05-01"), as.POSIXct("1969-05-02"), length=n_pts)
\donttest{
itimeplot(x, y, grp, chartOpts=list(xlab="Time"))
}
#'
}
\seealso{
\code{\link[=iplotCorr]{iplotCorr()}}, \code{\link[=iplotCurves]{iplotCurves()}}, \code{\link[=itriplot]{itriplot()}},
\code{\link[=idotplot]{idotplot()}}, \code{\link[=iplotPXG]{iplotPXG()}}
}
\keyword{hplot}
