% Generated by roxygen2 (4.0.0): do not edit by hand
\name{manyboxplots}
\alias{manyboxplots}
\title{Modern boxplot linked to underlying histrograms}
\usage{
manyboxplots(dat, qu = c(0.001, 0.01, 0.1, 0.25), orderByMedian = TRUE,
  breaks = 251, file, onefile = FALSE, openfile = TRUE,
  title = "Many box plots", legend)
}
\arguments{
  \item{dat}{Data matrix (individuals x variables)}

  \item{qu}{Quantiles to plot (All with 0 < qu < 0.5)}

  \item{orderByMedian}{If TRUE, reorder individuals by
  their median}

  \item{breaks}{Number of break points in the histogram}

  \item{file}{Optional character vector with file to
  contain the output}

  \item{onefile}{If TRUE, have output file contain all
  necessary javascript/css code}

  \item{openfile}{If TRUE, open the plot in the default web
  browser}

  \item{title}{Character string with title for plot}

  \item{legend}{Character vector with text for a legend (to
  be combined to one string with \code{\link[base]{paste}},
  with \code{collapse=''})}
}
\value{
Character string with the name of the file created.
}
\description{
Creates an interactive graph for a large set of box plots (rendered
as lines connecting the quantiles), linked to underlying histograms.
}
\examples{
n.ind <- 500
n.gene <- 10000
expr <- matrix(rnorm(n.ind * n.gene, (1:n.ind)/n.ind*3), ncol=n.gene)
dimnames(expr) <- list(paste0("ind", 1:n.ind),
                       paste0("gene", 1:n.gene))
manyboxplots(expr)
}

