% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print_qtlcharts_resources.R
\name{print_qtlcharts_resources}
\alias{print_qtlcharts_resources}
\title{Print javascript resources for qtlcharts}
\usage{
print_qtlcharts_resources(main, panels, charts, onefile = TRUE)
}
\arguments{
\item{main}{Vector of character strings with the main resources to use
If missing or \code{NULL}, the default is \code{c("d3", "d3tip",
"colorbrewer", "panelutil")}.}

\item{panels}{Vector of character strings with the names of panel function
to include. If missing or \code{NULL}, all panel functions are
included.}

\item{charts}{Vector of character strings with the names of charts
to include. If missing or \code{NULL}, all chart functions are
included.}

\item{onefile}{If \code{TRUE}, print the source for the selected
    resources; if \code{FALSE}, just print links to the resources.}
}
\value{
None.
}
\description{
Print javascript resources, for use with R/qtlcharts within an R
Markdown document
}
\details{
If \code{onefile=FALSE}, the links to the selected
    resources are printed.

    If \code{onefile=TRUE}, the javascript code for all selected
    resources are simply read into R with \code{\link[base]{readLines}}
    and then printed with \code{\link[base]{cat}}.

    This is to be included in a knitr code chunk (with
    \code{results="asis"}) in an R Markdown document, for use by
    later calls to the interactive plots, each with
    \code{print=TRUE}.
}
\examples{
\dontrun{print_qtlcharts_resources(panels=c("curvechart", "heatmap"), charts="iheatmap")}
}
\keyword{utilities}

