% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qtlchartsversion}
\alias{qtlchartsversion}
\title{print the installed version of R/qtlcharts}
\usage{
qtlchartsversion()
}
\value{
Character string with version number
}
\description{
print the installed version of R/qtlcharts
}
\examples{
qtlchartsversion()
}

