% Generated by roxygen2 (4.0.0): do not edit by hand
\name{corr_w_scatter}
\alias{corr_w_scatter}
\title{Image of correlation matrix with linked scatterplot}
\usage{
corr_w_scatter(dat, group, rows, cols, reorder = TRUE, corr = cor(dat, use =
  "pairwise.complete.obs"), file, onefile = FALSE, openfile = TRUE,
  title = "Correlation matrix with linked scatterplot", caption)
}
\arguments{
  \item{dat}{Data matrix (individuals x variables)}

  \item{group}{Option vector of groups of individuals
  (e.g., a genotype)}

  \item{rows}{Selected rows of the correlation matrix to
  include in the image. Ignored if \code{corr} is
  provided.}

  \item{cols}{Selected columns of the correlation matrix to
  include in the image. Ignored if \code{corr} is
  provided.}

  \item{reorder}{If TRUE, reorder the variables by
  clustering. Ignored if \code{corr} is provided as a
  subset of the overall correlation matrix}

  \item{corr}{Correlation matrix (optional).}

  \item{file}{Optional character vector with file to
  contain the output}

  \item{onefile}{If TRUE, have output file contain all
  necessary javascript/css code}

  \item{openfile}{If TRUE, open the plot in the default web
  browser}

  \item{title}{Character string with title for plot}

  \item{caption}{Character vector with text for a caption
  (to be combined to one string with
  \code{\link[base]{paste}}, with \code{collapse=''})}
}
\value{
Character string with the name of the file created.
}
\description{
Creates an interactive graph with an image of a
correlation matrix linked to underlying scatterplots.
}
\details{
\code{corr} may be provided as a subset of the overall
correlation matrix for the columns of \code{dat}. In this case, the
\code{reorder}, \code{rows} and \code{cols} arguments are ignored. The row and
column names of \code{corr} must match the names of some subset of
columns of \code{dat}.
}
\examples{
data(geneExpr)
corr_w_scatter(geneExpr$expr, geneExpr$genotype)
}
\seealso{
\code{\link{curves_w_2scatter}}
}
\keyword{hplot}

