% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{geneExpr}
\alias{geneExpr}
\title{Anonymized gene expression data}
\format{A list containing a matrix \code{expr} with the gene
expression data plus a vector \code{genotype} with the genotypes.}
\usage{
data(geneExpr)
}
\description{
An anonymized set of gene expression values, for 100
genes all influenced by a common locus, plus a vector of genotypes
for the 491 individuals.
}
\examples{
data(geneExpr)
iplotCorr(geneExpr$expr, geneExpr$genotype, reorder=TRUE,
          title="geneExpr correlations linked to scatterplots")
}
\author{
Karl W Broman, 2013-05-16
}
\keyword{datasets}

