% Generated by roxygen2 (4.0.0): do not edit by hand
\name{iheatmap}
\alias{iheatmap}
\title{Interactive heat map}
\usage{
iheatmap(z, x, y, file, onefile = FALSE, openfile = TRUE, title = "",
  caption, chartOpts = NULL)
}
\arguments{
  \item{x}{Vector of numeric values for the x-axis}

  \item{y}{Vector of numeric values for the y-axis}

  \item{z}{Numeric matrix of dim \code{length(x)} x
  \code{length(y)}}

  \item{file}{Optional character vector with file to
  contain the output}

  \item{onefile}{If TRUE, have output file contain all
  necessary javascript/css code}

  \item{openfile}{If TRUE, open the plot in the default web
  browser}

  \item{title}{Character string with title for plot}

  \item{caption}{Character vector with text for a caption
  (to be combined to one string with
  \code{\link[base]{paste}}, with \code{collapse=''})}

  \item{chartOpts}{A list of options for configuring the
  chart (see the coffeescript code). Each element must be
  named using the corresponding option.}
}
\value{
Character string with the name of the file created.
}
\description{
Creates an interactive heatmap, with each cell linked to
plots of horizontal and vertical slices
}
\details{
By default, the z-axis limits are from \code{-max(abs(z))}
to \code{max(abs(z))}, and negative cells are colored blue to white
which positive cells are colored white to red.
}
\examples{
n <- 101
x <- y <- seq(-pi, pi, len=n)
z <- matrix(ncol=n, nrow=n)
for(i in seq(along=x))
  for(j in seq(along=y))
    z[i,j] <- sin(x[i]) + cos(y[j])
iheatmap(z, x, y, title = "iheatmap example")
}
\seealso{
\code{\link{iplotCorr}}
}
\keyword{hplot}

