% Generated by roxygen2 (4.0.0): do not edit by hand
\name{iplotCorr}
\alias{iplotCorr}
\title{Image of correlation matrix with linked scatterplot}
\usage{
iplotCorr(mat, group, rows, cols, reorder = FALSE, corr = cor(mat, use =
  "pairwise.complete.obs"), file, onefile = FALSE, openfile = TRUE,
  title = "", chartdivid = "chart", caption, chartOpts = NULL,
  print = FALSE)
}
\arguments{
\item{mat}{Data matrix (individuals x variables)}

\item{group}{Optional vector of groups of individuals (e.g., a genotype)}

\item{rows}{Selected rows of the correlation matrix to include in
the image. Ignored if \code{corr} is provided.}

\item{cols}{Selected columns of the correlation matrix to include
in the image. Ignored if \code{corr} is provided.}

\item{reorder}{If TRUE, reorder the variables by
clustering. Ignored if \code{corr} is provided as a subset of the
overall correlation matrix}

\item{corr}{Correlation matrix (optional).}

\item{file}{Optional character vector with file to contain the
output}

\item{onefile}{If TRUE, have output file contain all necessary
javascript/css code}

\item{openfile}{If TRUE, open the plot in the default web browser}

\item{title}{Character string with title for plot}

\item{chartdivid}{Character string for id of div to hold the chart}

\item{caption}{Character vector with text for a caption (to be
combined to one string with \code{\link[base]{paste}}, with
\code{collapse=''})}

\item{chartOpts}{A list of options for configuring the chart (see
the coffeescript code). Each element must be named using the
corresponding option.}

\item{print}{If TRUE, print the output, rather than writing it to a file,
for use within an R Markdown document.}
}
\value{
Character string with the name of the file created.
}
\description{
Creates an interactive graph with an image of a
correlation matrix linked to underlying scatterplots.
}
\details{
\code{corr} may be provided as a subset of the overall
correlation matrix for the columns of \code{mat}. In this case, the
\code{reorder}, \code{rows} and \code{cols} arguments are ignored. The row and
column names of \code{corr} must match the names of some subset of
columns of \code{mat}.

Individual IDs are taken from \code{rownames(mat)}; they must match
\code{names(group)}.
}
\examples{
data(geneExpr)
iplotCorr(geneExpr$expr, geneExpr$genotype, reorder=TRUE,
               title = "iplotCorr example",
               chartOpts=list(cortitle="Correlation matrix",
                              scattitle="Scatterplot"))
}
\seealso{
\code{\link{iheatmap}}, \code{\link{iplotCurves}}
}
\keyword{hplot}

