% Generated by roxygen2 (4.0.0): do not edit by hand
\name{iplotCurves}
\alias{iplotCurves}
\title{Plot of a bunch of curves, linked to points in scatterplots}
\usage{
iplotCurves(curveMatrix, times, scatter1 = NULL, scatter2 = NULL,
  group = NULL, file, onefile = FALSE, openfile = TRUE, title = "",
  caption, chartOpts = NULL, ...)
}
\arguments{
  \item{curveMatrix}{Matrix (dim n_ind x n_times) with
  outcomes}

  \item{times}{Vector (length n_times) with time points for
  the columns of curveMatrix}

  \item{scatter1}{Matrix (dim n_ind x 2) with data for the
  first scatterplot}

  \item{scatter2}{Matrix (dim n_ind x 2) with data for the
  second scatterplot}

  \item{group}{Optional vector of groups of individuals
  (e.g., a genotype)}

  \item{file}{Optional character vector with file to
  contain the output}

  \item{onefile}{If TRUE, have output file contain all
  necessary javascript/css code}

  \item{openfile}{If TRUE, open the plot in the default web
  browser}

  \item{title}{Character string with title for plot}

  \item{caption}{Character vector with text for a caption
  (to be combined to one string with
  \code{\link[base]{paste}}, with \code{collapse=''})}

  \item{chartOpts}{A list of options for configuring the
  chart (see the coffeescript code). Each element must be
  named using the corresponding option.}

  \item{...}{Additional arguments passed to the
  \code{\link[RJSONIO]{toJSON}} function}
}
\value{
Character string with the name of the file created.
}
\description{
Creates an interactive graph with a panel having a number of curves
(say, a phenotype measured over time) linked to one or two (or no) scatter plots
(say, of the first vs middle and middle vs last times).
}
\examples{
# random growth curves, based on some data
times <- 1:16
n <- 100
start <- rnorm(n, 5.2, 0.8)
slope1to5 <- rnorm(n, 2.6, 0.5)
slope5to16 <- rnorm(n, 0.24 + 0.09*slope1to5, 0.195)
y <- matrix(ncol=16, nrow=n)
y[,1] <- start
for(j in 2:5)
  y[,j] <- y[,j-1] + slope1to5
for(j in 6:16)
  y[,j] <- y[,j-1] + slope5to16
y <- y + rnorm(prod(dim(y)), 0, 0.35)

# Make the plot
iplotCurves(y, times, y[,c(1,5)], y[,c(5,16)],
                 title = "iplotCurves example",
                 chartOpts=list(curves_xlab="Time", curves_ylab="Size",
                                scat1_xlab="Size at T=1", scat1_ylab="Size at T=5",
                                scat2_xlab="Size at T=5", scat2_ylab="Size at T=16"))
}
\seealso{
\code{\link{iplotCorr}}
}
\keyword{hplot}

