% Generated by roxygen2 (4.0.1): do not edit by hand
\name{iplotMap}
\alias{iplotMap}
\title{Interactive genetic map plot}
\usage{
iplotMap(map, shift = FALSE, file, onefile = FALSE, openfile = TRUE,
  title = "", chartdivid = "chart", caption, chartOpts = NULL,
  digits = 4, print = FALSE)
}
\arguments{
\item{map}{Object of class \code{"map"}, a list with each component
being a vector of marker positions.}

\item{shift}{If TRUE, shift each chromsome so that the initial marker
is at position 0.}

\item{file}{Optional character vector with file to contain the
output.}

\item{onefile}{If TRUE, have output file contain all necessary
javascript/css code.}

\item{openfile}{If TRUE, open the plot in the default web browser.}

\item{title}{Character string with title for plot.}

\item{chartdivid}{Character string for id of div to hold the chart}

\item{caption}{Character vector with text for a caption (to be
combined to one string with \code{\link[base]{paste}}, with
\code{collapse=''})}

\item{chartOpts}{A list of options for configuring the chart.  Each
element must be named using the corresponding option. See details.}

\item{digits}{Number of digits in JSON; passed to \cite{\link[jsonlite]{toJSON}}.}

\item{print}{If TRUE, print the output, rather than writing it to a file,
for use within an R Markdown document.}
}
\value{
Character string with the name of the file created.
}
\description{
Creates an interactive graph of a genetic marker map.
}
\examples{
data(hyper)
map <- pull.map(hyper)
\donttest{
# open iplotMap in web browser
iplotMap(map, shift=TRUE, title="iplotMap example")}
\dontshow{
# save to temporary file but don't open
iplotMap(map, shift=TRUE, title="iplotMap example",
         openfile=FALSE)}
}
\seealso{
\code{\link{iplotScanone}}, \code{\link{iplotPXG}}
}
\keyword{hplot}

