% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/iplotMScanone.R
\name{iplotMScanone}
\alias{iplotMScanone}
\title{Interactive LOD curve}
\usage{
iplotMScanone(scanoneOutput, cross, lodcolumn, pheno.col, times = NULL,
  effects, chr, file, onefile = FALSE, openfile = TRUE, title = "",
  chartdivid = "chart", caption, chartOpts = NULL, digits = 4,
  print = FALSE)
}
\arguments{
\item{scanoneOutput}{Object of class \code{"scanone"}, as output
from \code{\link[qtl]{scanone}}.}

\item{cross}{(Optional) Object of class \code{"cross"}, see
\code{\link[qtl]{read.cross}}.}

\item{lodcolumn}{Numeric value indicating LOD score column to plot.}

\item{pheno.col}{(Optional) Phenotype column in cross object.}

\item{times}{(Optional) Vector (length equal to the number of LOD
score columns) with quantitative values to which the different LOD
score columns correspond (times of measurements, or something like
age or dose).  These need to be ordered and equally-spaced. If
omitted, the names of the columns in \code{scanoneOutput} are used
and treated as qualitative.}

\item{effects}{(Optional)}

\item{chr}{(Optional) Optional vector indicating the chromosomes
for which LOD scores should be calculated. This should be a vector
of character strings referring to chromosomes by name; numeric
values are converted to strings. Refer to chromosomes with a
preceding - to have all chromosomes but those considered. A logical
(TRUE/FALSE) vector may also be used.}

\item{file}{Optional character vector with file to contain the
output}

\item{onefile}{If TRUE, have output file contain all necessary
javascript/css code}

\item{openfile}{If TRUE, open the plot in the default web browser}

\item{title}{Character string with title for plot}

\item{chartdivid}{Character string for id of div to hold the chart}

\item{caption}{Character vector with text for a caption (to be
combined to one string with \code{\link[base]{paste}}, with
\code{collapse=""})}

\item{chartOpts}{A list of options for configuring the chart (see
the coffeescript code). Each element must be named using the
corresponding option.}

\item{digits}{Number of digits in JSON; passed to
\code{\link[jsonlite]{toJSON}}}

\item{print}{If TRUE, print the output, rather than writing it to a file,
for use within an R Markdown document.}
}
\value{
Character string with the name of the file created.
}
\description{
Creates an interactive graph of a set of single-QTL genome scans, as
calculated by \code{\link[qtl]{scanone}}. If \code{cross} or
\code{effects} are provide, LOD curves will be linked to a panel
with estimated QTL effects.
}
\details{
If \code{cross} is provided, Haley-Knott regression is
used to estimate QTL effects at each pseudomarker.
}
\examples{
data(grav)
grav <- calc.genoprob(grav, step=1)
grav <- reduce2grid(grav)

# we're going to subset the phenotypes
phecol <- seq(1, nphe(grav), by=5)

# the times were saved as an attributed
times <- attr(grav, "time")[phecol]

# genome scan
out <- scanone(grav, phe=phecol, method="hk")

\donttest{
# plot with qualitative labels on y-axis (in web browser)
iplotMScanone(out, title="iplotMScanone example, no effects")}
\dontshow{
# save to temporary file but don't open
iplotMScanone(out, title="iplotMScanone example, no effects",
              openfile=FALSE)}

\donttest{
# plot with quantitative y-axis
iplotMScanone(out, times=times, title="iplotMScanone example, no effects")}
\dontshow{
# save to temporary file but don't open
iplotMScanone(out, times=times, title="iplotMScanone example, no effects",
              openfile=FALSE)}

# estimate QTL effect for each time point at each genomic position
eff <- estQTLeffects(grav, phe=seq(1, nphe(grav), by=5), what="effects")

\donttest{
# plot with QTL effects included (and with quantitative y-axis)
iplotMScanone(out, effects=eff, times=times,
              title="iplotMScanone example, with effects",
              chartOpts=list(eff_ylab="QTL effect", eff_xlab="Time (hrs)"))}
\dontshow{
# save to temporary file but don't open
iplotMScanone(out, effects=eff, times=times,
              title="iplotMScanone example, with effects",
              chartOpts=list(eff_ylab="QTL effect", eff_xlab="Time (hrs)"),
              openfile=FALSE)}
}
\seealso{
\code{\link{iplotScanone}}
}
\keyword{hplot}

