% Generated by roxygen2 (4.0.0): do not edit by hand
\name{iplotMScanone}
\alias{iplotMScanone}
\title{Interactive LOD curve}
\usage{
iplotMScanone(scanoneOutput, cross, lodcolumn, pheno.col, effects, chr, file,
  onefile = FALSE, openfile = TRUE, title = "", caption,
  chartOpts = NULL, ...)
}
\arguments{
  \item{scanoneOutput}{Object of class \code{"scanone"}, as
  output from \code{\link[qtl]{scanone}}.}

  \item{cross}{(Optional) Object of class \code{"cross"},
  see \code{\link[qtl]{read.cross}}.}

  \item{lodcolumn}{Numeric value indicating LOD score
  column to plot.}

  \item{pheno.col}{(Optional) Phenotype column in cross
  object.}

  \item{effects}{(Optional)}

  \item{chr}{(Optional) Optional vector indicating the
  chromosomes for which LOD scores should be calculated.
  This should be a vector of character strings referring to
  chromosomes by name; numeric values are converted to
  strings. Refer to chromosomes with a preceding - to have
  all chromosomes but those considered. A logical
  (TRUE/FALSE) vector may also be used.}

  \item{file}{Optional character vector with file to
  contain the output}

  \item{onefile}{If TRUE, have output file contain all
  necessary javascript/css code}

  \item{openfile}{If TRUE, open the plot in the default web
  browser}

  \item{title}{Character string with title for plot}

  \item{caption}{Character vector with text for a caption
  (to be combined to one string with
  \code{\link[base]{paste}}, with \code{collapse=''})}

  \item{chartOpts}{A list of options for configuring the
  chart (see the coffeescript code). Each element must be
  named using the corresponding option.}

  \item{...}{Additional arguments passed to the
  \code{\link[RJSONIO]{toJSON}} function}
}
\value{
Character string with the name of the file created.
}
\description{
Creates an interactive graph of a set of single-QTL genome scans, as
calculated by \code{\link[qtl]{scanone}}. If \code{cross} or
\code{effects} are provide, LOD curves will be linked to a panel
with estimated QTL effects.
}
\details{
If \code{cross} is provided, Haley-Knott regression is
used to estimate QTL effects at each pseudomarker.
}
\examples{
data(grav)
grav <- calc.genoprob(grav, step=1)
grav <- reduce2grid(grav)
out <- scanone(grav, phe=seq(1, nphe(grav), by=5))
iplotMScanone(out, title="iplotMScanone example")
}
\seealso{
\code{\link{iplotScanone}}
}
\keyword{hplot}

