% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iplotPXG.R
\name{iplotPXG}
\alias{iplotPXG}
\title{Interactive phenotype x genotype plot}
\usage{
iplotPXG(cross, marker, pheno.col = 1, chartOpts = NULL,
  fillgenoArgs = NULL)
}
\arguments{
\item{cross}{Object of class \code{"cross"}, see
\code{\link[qtl]{read.cross}}.}

\item{marker}{Character string with marker name.}

\item{pheno.col}{Phenotype column in cross object.}

\item{chartOpts}{A list of options for configuring the chart.  Each
element must be named using the corresponding option.}

\item{fillgenoArgs}{List of named arguments to pass to
  \code{\link[qtl]{fill.geno}}, if needed.}
}
\value{
None.
}
\description{
Creates an interactive graph of phenotypes vs genotypes at a marker.
}
\details{
The function \code{\link[qtl]{fill.geno}} is used to
impute missing genotypes, with arguments passed as a list, for
example \code{fillgenoArgs=list(method="argmax", error.prob=0.002,
map.function="c-f")}.

Individual IDs (viewable when hovering over a point) are taken from
the input \code{cross} object, using the \code{\link[qtl]{getid}}
function in R/qtl.
}
\examples{
library(qtl)
data(hyper)
marker <- sample(markernames(hyper), 1)
\donttest{
iplotPXG(hyper, marker, chartOpts=list(width=500))}
}
\seealso{
\code{\link{iplotScanone}}, \code{\link{iplotMap}}
}
\keyword{hplot}

