% Generated by roxygen2 (4.0.0): do not edit by hand
\name{iplotPXG}
\alias{iplotPXG}
\title{Interactive phenotype x genotype plot}
\usage{
iplotPXG(cross, marker, pheno.col = 1, file, onefile = FALSE,
  openfile = TRUE, title = "", caption, chartOpts = list(title =
  marker[1]), fillgenoArgs = NULL, ...)
}
\arguments{
\item{cross}{Object of class \code{"cross"}, see
\code{\link[qtl]{read.cross}}.}

\item{marker}{Character string with marker name.}

\item{pheno.col}{Phenotype column in cross object.}

\item{file}{Optional character vector with file to contain the
output.}

\item{onefile}{If TRUE, have output file contain all necessary
javascript/css code.}

\item{openfile}{If TRUE, open the plot in the default web browser.}

\item{title}{Character string with title for plot.}

\item{caption}{Character vector with text for a caption (to be
combined to one string with \code{\link[base]{paste}}, with
\code{collapse=''})}

\item{chartOpts}{A list of options for configuring the chart.  Each
element must be named using the corresponding option. See details.}

\item{fillgenoArgs}{List of named arguments to pass to
\code{\link[qtl]{fill.geno}}, if needed.}

\item{...}{Passed to \cite{\link[jsonlite]{toJSON}}.}
}
\value{
Character string with the name of the file created.
}
\description{
Creates an interactive graph of phenotypes vs genotypes at a marker.
}
\details{
The function \code{\link[qtl]{fill.geno}} is used to
impute missing genotypes, with arguments passed as a list, for
example \code{fillgenoArgs=list(method="argmax", error.prob=0.002,
map.function="c-f")}.

Individual IDs (viewable when hovering over a point) are taken from
the input \code{cross} object, using the \code{\link[qtl]{getid}}
function in R/qtl.
}
\examples{
data(hyper)
marker <- sample(markernames(hyper), 1)
iplotPXG(hyper, marker, title="iplotPXG example")
}
\seealso{
\code{\link{iplotScanone}}, \code{\link{iplotMap}}
}
\keyword{hplot}

