% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iplotRF.R
\name{iplotRF}
\alias{iplotRF}
\title{Interactive plot of recombination fractions}
\usage{
iplotRF(cross, chr, chartOpts = NULL)
}
\arguments{
\item{cross}{Object of class \code{"cross"}, see
\code{\link[qtl]{read.cross}}.}

\item{chr}{Optional vector indicating chromosomes to include. This
should be a vector of character strings referring to chromosomes by
name; numeric values are converted to strings.  Refer to
chromosomes with a preceding \code{-} to have all chromosomes but
those considered.  A logical (TRUE/FALSE) vector may also be used.}

\item{chartOpts}{A list of options for configuring the chart.  Each
    element must be named using the corresponding option.}
}
\value{
None.
}
\description{
Creates an interactive graph of estimated recombination fractions and LOD scores for all pairs of markers.
}
\examples{
library(qtl)
data(hyper)
hyper <- est.rf(hyper)
\donttest{
iplotRF(hyper)}
}
\seealso{
\code{\link[qtl]{est.rf}}, \code{\link[qtl]{plotRF}}
}
\keyword{hplot}

