% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/iplotScantwo.R
\name{iplotScantwo}
\alias{iplotScantwo}
\title{Interactive plot of 2d genome scan}
\usage{
iplotScantwo(scantwoOutput, cross, lodcolumn = 1, pheno.col = 1, chr, file,
  onefile = FALSE, openfile = TRUE, title = "", chartdivid = "chart",
  caption, chartOpts = NULL, digits = 4, print = FALSE)
}
\arguments{
\item{scantwoOutput}{Output of \code{\link[qtl]{scantwo}}}

\item{cross}{(Optional) Object of class \code{"cross"}, see
\code{\link[qtl]{read.cross}}.}

\item{lodcolumn}{Numeric value indicating LOD score column to plot.}

\item{pheno.col}{(Optional) Phenotype column in cross object.}

\item{chr}{(Optional) Optional vector indicating the chromosomes
for which LOD scores should be calculated. This should be a vector
of character strings referring to chromosomes by name; numeric
values are converted to strings. Refer to chromosomes with a
preceding - to have all chromosomes but those considered. A logical
(TRUE/FALSE) vector may also be used.}

\item{file}{Optional character vector with file to contain the
output.}

\item{onefile}{If TRUE, have output file contain all necessary
javascript/css code.}

\item{openfile}{If TRUE, open the plot in the default web browser.}

\item{title}{Character string with title for plot.}

\item{chartdivid}{Character string for id of div to hold the chart}

\item{caption}{Character vector with text for a caption (to be
combined to one string with \code{\link[base]{paste}}, with
\code{collapse=""})}

\item{chartOpts}{A list of options for configuring the chart.  Each
element must be named using the corresponding option. See details.}

\item{digits}{Number of digits in JSON; passed to \cite{\link[jsonlite]{toJSON}}.}

\item{print}{If TRUE, print the output, rather than writing it to a file,
for use within an R Markdown document.}
}
\value{
Character string with the name of the file created.
}
\description{
Creates an interactive plot of the results of
\code{\link[qtl]{scantwo}}, for a two-dimensional, two-QTL genome
scan.
}
\examples{
library(qtl)
data(fake.f2)
\dontshow{fake.f2 <- fake.f2[c(1, 13, "X"),]}
fake.f2 <- calc.genoprob(fake.f2, step=5)
out <- scantwo(fake.f2, method="hk", verbose=FALSE)
\donttest{
# open iplotScantwo in web browser
iplotScantwo(out, fake.f2)}
\dontshow{
# save to temporary file but don't open
iplotScantwo(out, fake.f2, openfile=FALSE, chartOpts=list(zthresh=2))}
}
\seealso{
\code{\link{iplotScanone}}
}
\keyword{hplot}

