% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlcharts.R
\docType{package}
\name{qtlcharts-package}
\alias{qtlcharts-package}
\alias{qtlcharts}
\title{R/qtlcharts: Interactive charts for QTL data}
\description{
A QTL is a \emph{quantitative trait locus}: a genetic locus that
contributes to variation in a quantitative trait. The goal of
\href{https://kbroman.org/qtlcharts/}{R/qtlcharts} is to provide
interactive data visualizations for QTL analyses, and to make these
visualizations available from R. It is a companion to the
\href{https://rqtl.org}{R/qtl} package.
}
\details{
Vignettes online at the \href{https://kbroman.org/qtlcharts/}{R/qtlcharts website}:
\itemize{
\item \href{https://kbroman.org/qtlcharts/assets/vignettes/userGuide.html}{User guide}
\item \href{https://kbroman.org/qtlcharts/assets/vignettes/develGuide.html}{Developer guide}
\item \href{https://kbroman.org/qtlcharts/assets/vignettes/Rmarkdown.html}{Use with R Markdown}
[\href{https://github.com/kbroman/qtlcharts/blob/gh-pages/assets/vignettes/Rmarkdown.Rmd}{Rmd source}]
\item \href{https://kbroman.org/qtlcharts/assets/vignettes/chartOpts.html}{List of chart customization options}
\item \href{https://github.com/kbroman/shiny_qtlcharts}{Examples of use with Shiny}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kbroman.org/qtlcharts/}
  \item \url{https://github.com/kbroman/qtlcharts}
  \item Report bugs at \url{https://github.com/kbroman/qtlcharts/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

Other contributors:
\itemize{
  \item Michael Bostock (d3.js library in htmlwidgets/lib, https://d3js.org) [contributor, copyright holder]
  \item jQuery Foundation (jQuery library in htmlwidgets/lib, https://jquery.com) [copyright holder]
  \item jQuery contributors (jQuery library in htmlwidgets/lib; see https://github.com/jquery/jquery/blob/master/AUTHORS.txt) [contributor]
  \item jQuery UI contributors (jQuery UI library in htmlwidgets/lib; see https://jqueryui.com/about/) [contributor]
}

}
