% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qtlcharts.R
\docType{package}
\name{qtlcharts-package}
\alias{qtlcharts}
\alias{qtlcharts-package}
\title{Interactive charts for QTL data}
\description{
A QTL is a \emph{quantitative trait locus}: a genetic locus that
contributes to variation in a quantitative trait. The goal of
\href{http://kbroman.org/qtlcharts}{R/qtlcharts} is to provide
interactive data visualizations for QTL analyses, and to make these
visualizations available from R.
}
\details{
See \url{http://kbroman.org/qtlcharts}.
}

