% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlcharts.R
\docType{package}
\name{qtlcharts-package}
\alias{qtlcharts}
\alias{qtlcharts-package}
\title{R/qtlcharts: Interactive charts for QTL data}
\description{
A QTL is a \emph{quantitative trait locus}: a genetic locus that
contributes to variation in a quantitative trait. The goal of
\href{http://kbroman.org/qtlcharts}{R/qtlcharts} is to provide
interactive data visualizations for QTL analyses, and to make these
visualizations available from R. It is a companion to the
\href{http://www.rqtl.org}{R/qtl} package.
}
\details{
Vignettes online at the \href{http://kbroman.org/qtlcharts}{R/qtlcharts website}:

\itemize{
\item \href{http://kbroman.org/qtlcharts/assets/vignettes/userGuide.html}{User guide}
\item \href{http://kbroman.org/qtlcharts/assets/vignettes/develGuide.html}{Developer guide}
\item \href{http://kbroman.org/qtlcharts/assets/vignettes/Rmarkdown.html}{Use with R Markdown}
  [\href{https://github.com/kbroman/qtlcharts/blob/gh-pages/assets/vignettes/Rmarkdown.Rmd}{Rmd source}]
\item \href{http://kbroman.org/qtlcharts/assets/vignettes/chartOpts.html}{List of chart customization options}
}
}

