% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iboxplot.R, R/idotplot.R, R/iheatmap.R,
%   R/ipleiotropy.R, R/iplot.R, R/iplotCorr.R, R/iplotCurves.R,
%   R/iplotMScanone.R, R/iplotMap.R, R/iplotRF.R, R/iplotScanone.R,
%   R/iplotScantwo.R, R/itriplot.R, R/qtlcharts.R, R/scat2scat.R
\name{iboxplot_output}
\alias{iboxplot_output}
\alias{iboxplot_render}
\alias{idotplot_output}
\alias{idotplot_render}
\alias{iheatmap_output}
\alias{iheatmap_render}
\alias{ipleiotropy_output}
\alias{ipleiotropy_render}
\alias{iplot_output}
\alias{iplot_render}
\alias{iplotCorr_output}
\alias{iplotCorr_render}
\alias{iplotCurves_output}
\alias{iplotCurves_render}
\alias{iplotMScanone_output}
\alias{iplotMScanone_render}
\alias{iplotMap_output}
\alias{iplotMap_render}
\alias{iplotRF_output}
\alias{iplotRF_render}
\alias{iplotScanone_output}
\alias{iplotScanone_render}
\alias{iplotScantwo_output}
\alias{iplotScantwo_render}
\alias{itriplot_output}
\alias{itriplot_render}
\alias{qtlcharts-shiny}
\alias{scat2scat_output}
\alias{scat2scat_render}
\title{Shiny bindings for R/qtlcharts widgets}
\usage{
iboxplot_output(outputId, width = "100\%", height = "900")

iboxplot_render(expr, env = parent.frame(), quoted = FALSE)

idotplot_output(outputId, width = "100\%", height = "530")

idotplot_render(expr, env = parent.frame(), quoted = FALSE)

iheatmap_output(outputId, width = "100\%", height = "1000")

iheatmap_render(expr, env = parent.frame(), quoted = FALSE)

ipleiotropy_output(outputId, width = "100\%", height = "580")

ipleiotropy_render(expr, env = parent.frame(), quoted = FALSE)

iplot_output(outputId, width = "100\%", height = "580")

iplot_render(expr, env = parent.frame(), quoted = FALSE)

iplotCorr_output(outputId, width = "100\%", height = "1000")

iplotCorr_render(expr, env = parent.frame(), quoted = FALSE)

iplotCurves_output(outputId, width = "100\%", height = "1000")

iplotCurves_render(expr, env = parent.frame(), quoted = FALSE)

iplotMScanone_output(outputId, width = "100\%", height = "580")

iplotMScanone_render(expr, env = parent.frame(), quoted = FALSE)

iplotMap_output(outputId, width = "100\%", height = "680")

iplotMap_render(expr, env = parent.frame(), quoted = FALSE)

iplotRF_output(outputId, width = "100\%", height = "1000")

iplotRF_render(expr, env = parent.frame(), quoted = FALSE)

iplotScanone_output(outputId, width = "100\%", height = "580")

iplotScanone_render(expr, env = parent.frame(), quoted = FALSE)

iplotScantwo_output(outputId, width = "100\%", height = "1000")

iplotScantwo_render(expr, env = parent.frame(), quoted = FALSE)

itriplot_output(outputId, width = "100\%", height = "530")

itriplot_render(expr, env = parent.frame(), quoted = FALSE)

scat2scat_output(outputId, width = "100\%", height = "530")

scat2scat_render(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like `"100\%"`,
`"400px"`, `"auto"`) or a number, which will be coerced to a
string and have `"px"` appended.}

\item{expr}{An expression that generates a networkD3 graph}

\item{env}{The environment in which to evaluate `expr`.}

\item{quoted}{Is `expr` a quoted expression (with `quote()`)? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using R/qtlcharts widgets within Shiny
applications and interactive Rmd documents.
}
