% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qtlchartsversion.R
\name{qtlchartsversion}
\alias{qtlchartsversion}
\title{print the installed version of R/qtlcharts}
\usage{
qtlchartsversion()
}
\value{
Character string with version number
}
\description{
print the installed version of R/qtlcharts
}
\examples{
qtlchartsversion()
}

